+++ services/workflows-service/src/env.ts
@@ -86,9 +86,10 @@ export const serverEnvSchema = {
     .default('/dev/customers/')
     .describe('AWS Secrets Manager prefix'),
 
-  IN_MEMORY_SECRET_ACQUIRER_ID: z.string().optional(),
-  IN_MEMORY_SECRET_PRIVATE_KEY: z.string().optional(),
-  IN_MEMORY_SECRET_CONSUMER_KEY: z.string().optional(),
+  // IN_MEMORY is reserved for environment variables
+  IN_MEMORIES_SECRET_ACQUIRER_ID: z.string().optional(),
+  IN_MEMORIES_SECRET_PRIVATE_KEY: z.string().optional(),
+  IN_MEMORIES_SECRET_CONSUMER_KEY: z.string().optional(),
 };
 
 if (!process.env['ENVIRONMENT_NAME'] || process.env['ENVIRONMENT_NAME'] === 'local') {