+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/headless-example
 
+## 0.3.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.26
+  - @ballerine/workflow-browser-sdk@0.6.37
+
 ## 0.3.35
 
 ### Patch Changes